more init fn renames
