+++ src/animation/skeletal.ts
@@ -20,7 +20,7 @@ export const PoseDef = EM.defineComponent("pose", (current?: number) => ({
   repeat: [] as QueuedAnimation[],
 }));
 
-export function registerSkeletalAnimSystems() {
+export function initSkeletalAnimSystems() {
   EM.addSystem(
     "pose",
     Phase.RENDER_PRE_DRAW,
