+++ src/render/motion-smoothing.ts
@@ -9,17 +9,28 @@ import {
   copyFrame,
   updateFrameFromPosRotScale,
   updateFrameFromTransform,
+  Frame,
 } from "../physics/transform.js";
 import { computeNewError, reduceError } from "../utils/smoothing.js";
 import { RemoteUpdatesDef } from "../net/components.js";
 import { Phase } from "../ecs/sys-phase.js";
-import { RenderableDef } from "./renderer-ecs.js";
+import { RenderableDef, RendererWorldFrameDef } from "./renderer-ecs.js";
 import { DONT_SMOOTH_WORLD_FRAME } from "../flags.js";
 import { DeletedDef } from "../ecs/delete.js";
-import { createFrame } from "../physics/nonintersection.js";
+import { WorldFrameDef, createFrame } from "../physics/nonintersection.js";
+import { tempVec3 } from "../matrix/temp-pool.js";
 
 // Determined via binary search--smaller -> jerky, larger -> floaty
 const ERROR_SMOOTHING_FACTOR = 0.75 ** (60 / 1000);
+
+const BLEND_SIMULATION_FRAMES_STRATEGY: "interpolate" | "extrapolate" | "none" =
+  "none";
+
+let _simulationAlpha = 0.0;
+
+export function setSimulationAlpha(to: number) {
+  _simulationAlpha = to;
+}
 
 export const MotionSmoothingDef = EM.defineComponent("motionSmoothing", () => {
   return {
@@ -182,4 +193,97 @@ export function initMotionSmoothingSystems(em: EntityManager) {
       }
     }
   );
+
+  em.addSystem(
+    "updateRendererWorldFrames",
+    Phase.RENDER_WORLDFRAMES,
+    [SmoothedWorldFrameDef, PrevSmoothedWorldFrameDef],
+    [],
+    (objs) => {
+      for (let o of objs) {
+        if (DONT_SMOOTH_WORLD_FRAME) {
+          // TODO(@darzu): HACK!
+          if (WorldFrameDef.isOn(o)) {
+            em.ensureComponentOn(o, RendererWorldFrameDef);
+            copyFrame(o.rendererWorldFrame, o.world);
+            // (o as any).rendererWorldFrame = o.world;
+          }
+          continue;
+        }
+
+        em.ensureComponentOn(o, RendererWorldFrameDef);
+
+        switch (BLEND_SIMULATION_FRAMES_STRATEGY) {
+          case "interpolate":
+            interpolateFrames(
+              _simulationAlpha,
+              o.rendererWorldFrame,
+              o.prevSmoothedWorldFrame,
+              o.smoothedWorldFrame
+            );
+            break;
+          case "extrapolate":
+            extrapolateFrames(
+              _simulationAlpha,
+              o.rendererWorldFrame,
+              o.prevSmoothedWorldFrame,
+              o.smoothedWorldFrame
+            );
+            break;
+          default:
+            copyFrame(o.rendererWorldFrame, o.smoothedWorldFrame);
+        }
+      }
+    }
+  );
+}
+
+function interpolateFrames(
+  alpha: number,
+  out: Frame,
+  prev: Frame,
+  next: Frame
+) {
+  vec3.lerp(prev.position, next.position, alpha, out.position);
+  quat.slerp(prev.rotation, next.rotation, alpha, out.rotation);
+  vec3.lerp(prev.scale, next.scale, alpha, out.scale);
+  updateFrameFromPosRotScale(out);
+}
+
+function extrapolateFrames(
+  alpha: number,
+  out: Frame,
+  prev: Frame,
+  next: Frame
+) {
+  // out.position = next.position + alpha * (next.position - prev.position)
+  // out.position = next.position + alpha * (next.position - prev.position)
+  vec3.sub(next.position, prev.position, out.position);
+  vec3.scale(out.position, alpha, out.position);
+  vec3.add(out.position, next.position, out.position);
+
+  // see https://answers.unity.com/questions/168779/extrapolating-quaternion-rotation.html
+  // see https://answers.unity.com/questions/168779/extrapolating-quaternion-rotation.html
+  quat.invert(prev.rotation, out.rotation);
+  quat.mul(next.rotation, out.rotation, out.rotation);
+  const axis = tempVec3();
+  let angle = quat.getAxisAngle(out.rotation, axis);
+  // ensure we take the shortest path
+  if (angle > Math.PI) {
+    angle -= Math.PI * 2;
+  }
+  if (angle < -Math.PI) {
+    angle += Math.PI * 2;
+  }
+  angle = angle * alpha;
+  quat.setAxisAngle(axis, angle, out.rotation);
+  quat.mul(out.rotation, next.rotation, out.rotation);
+
+  // out.scale = next.scale + alpha * (next.scale - prev.scale)
+  // out.scale = next.scale + alpha * (next.scale - prev.scale)
+  vec3.sub(next.scale, prev.scale, out.scale);
+  vec3.scale(out.scale, alpha, out.scale);
+  vec3.add(out.scale, next.scale, out.scale);
+
+  updateFrameFromPosRotScale(out);
 }
